feat(invoice): Send email to users
The email has a link to see a public version of the invoice.
There's a pdf version of the invoice attached on the email.

A job was created to generate a temporary pdf file to be used for the mailer to deliver it to all the linked emails present on the invoice.